+++ src/recorder.ts
@@ -33,8 +33,8 @@ export default class Recorder implements IRecorder {
 	private process: ChildProcessWithoutNullStreams | null = null;
 	private eventEmitter: EventEmitter;
 
-	constructor (private uri: string, private path: string, options: Options = {}) {
-		const errors = verifyAllOptions(path, options);
+	constructor (private uri: string, private destination: string, options: Options = {}) {
+		const errors = verifyAllOptions(destination, options);
 		if (errors.length) {
 			throw new RecorderValidationError('Options invalid', errors);
 		}
@@ -114,7 +114,7 @@ export default class Recorder implements IRecorder {
 				{
 					detached: false,
 					shell: true,
-					cwd: this.path,
+					cwd: this.destination,
 				},
 			);
 
@@ -165,15 +165,15 @@ export default class Recorder implements IRecorder {
 		const playlist = this.matchStarted(message);
 		if (playlist) {
 			this.eventEmitter.emit(Events.STARTED, {
-				path: this.path,
 				uri: this.uri,
-				segmentTime: this.segmentTime,
-				filePattern: this.filePattern,
-				dirSizeThreshold: this.dirSizeThreshold,
+				destination: this.destination,
+				playlist,
 				title: this.title,
+				filePattern: this.filePattern,
+				segmentTime: this.segmentTime,
+				dirSizeThreshold: this.dirSizeThreshold,
 				noAudio: this.noAudio,
 				ffmpegBinary: this.ffmpegBinary,
-				playlist,
 			});
 		}
 
@@ -188,7 +188,7 @@ export default class Recorder implements IRecorder {
 			if (!this.dirSizeThreshold) {
 				return;
 			}
-			const used = dirSize(this.path);
+			const used = dirSize(this.destination);
 			if (Math.ceil(used + used * APPROXIMATION_PERCENTAGE / 100) > this.dirSizeThreshold) {
 				this.eventEmitter.emit(Events.SPACE_FULL, {
 					threshold: this.dirSizeThreshold,