+++ test/events/file_created.spec.ts
@@ -1,7 +1,7 @@
 import { ChildProcessWithoutNullStreams } from 'child_process';
 import { mocked } from 'ts-jest/utils';
 import { verifyAllOptions } from '../../src/validators';
-import {mockSpawnProcess, URI, PATH} from '../test.helpers';
+import {mockSpawnProcess, URI, DESTINATION} from '../test.helpers';
 import Recorder, { RecorderEvents } from '../../src/recorder';
 
 jest.mock('../../src/validators');
@@ -16,7 +16,7 @@ beforeEach(() => {
 });
 
 test('should return filename if ffmpeg says: "Opening \'*.mp4\' for writing"', () => {
-	new Recorder(URI, PATH)
+	new Recorder(URI, DESTINATION)
 		.on(RecorderEvents.FILE_CREATED, eventHandler)
 		.start();
 
@@ -27,7 +27,7 @@ test('should return filename if ffmpeg says: "Opening \'*.mp4\' for writing"', (
 });
 
 test('should not handle event if ffmpeg says: "Opening \'*.m3u8.tmp\' for writing"', () => {
-	new Recorder(URI, PATH)
+	new Recorder(URI, DESTINATION)
 		.on(RecorderEvents.FILE_CREATED, eventHandler)
 		.start();