+++ test/events/space_full.spec.ts
@@ -2,7 +2,7 @@ import { ChildProcessWithoutNullStreams } from 'child_process';
 import { transformDirSizeThreshold, dirSize } from '../../src/helpers';
 import { mocked } from 'ts-jest/utils';
 import { verifyAllOptions } from '../../src/validators';
-import {mockSpawnProcess, URI, PATH} from '../test.helpers';
+import {mockSpawnProcess, URI, DESTINATION} from '../test.helpers';
 import Recorder, { RecorderEvents, RecorderError } from '../../src/recorder';
 
 jest.mock('../../src/validators');
@@ -20,7 +20,7 @@ beforeEach(() => {
 test('should not evaluate space if "threshold" is undefined', () => {
 	mocked(dirSize).mockReturnValue(Infinity);
 
-	new Recorder(URI, PATH)
+	new Recorder(URI, DESTINATION)
 		.on(RecorderEvents.SPACE_FULL, onSpaceFull)
 		.start();
 
@@ -36,7 +36,7 @@ test('should evaluate space and rise an event if "used" is close to the "thresho
 	mocked(dirSize).mockReturnValue(496);
 	const onStop = jest.fn().mockName('onStop');
 
-	new Recorder(URI, PATH, { dirSizeThreshold })
+	new Recorder(URI, DESTINATION, { dirSizeThreshold })
 		.on(RecorderEvents.SPACE_FULL, onSpaceFull)
 		.on(RecorderEvents.STOP, onStop)
 		.start();
@@ -58,7 +58,7 @@ test('should evaluate space but do not rise an event if "used" is far from the "
 	mocked(transformDirSizeThreshold).mockReturnValue(dirSizeThreshold);
 	mocked(dirSize).mockReturnValue(400);
 
-	new Recorder(URI, PATH, { dirSizeThreshold })
+	new Recorder(URI, DESTINATION, { dirSizeThreshold })
 		.on(RecorderEvents.SPACE_FULL, onSpaceFull)
 		.start();
 
@@ -76,7 +76,7 @@ test('should return RecorderError - space evaluation failed', () => {
 	});
 	const onError = jest.fn().mockName('onError');
 
-	new Recorder(URI, PATH, { dirSizeThreshold })
+	new Recorder(URI, DESTINATION, { dirSizeThreshold })
 		.on(RecorderEvents.SPACE_FULL, onSpaceFull)
 		.on(RecorderEvents.ERROR, onError)
 		.start();